Model_Drift_Detection_Automation/requirements.txt
scikit-learn
pandas
alibi-detect
apache-airflow==3.0.2